Attempt to modernize bar display on Win10
The current Bar has an intermittent bug on Win10 that causes it to be displayed
as a small 2-3px size square in the upper left corner. It could be related to
dpi changes in Win10 for Forms apps.

https://msdn.microsoft.com/library/windows/desktop/mt843498%28v=vs.85%29.aspx%28d=robot%29?f=255&MSPPError=-2147217396

This attempts to update the bar defaults to work nicely on Windows 10. Since the
bar is a WinForms app the dpi changes in Win10 may affect it:

In minimal testing it seems to solve the issue, further testing is still needed